Revert "Work on replacing Tank model."
This reverts commit 8d5797589ddfeadada297feceaf1320533559e46.

Today's attempts at replacing the non-free tank model have not gone to plan. Trying to build a new tank model based off a model from BlendSwap did not work out, the mesh was too complex and the coordinate system didn't match. My limited knowledge of Blender and the Unity Asset pipeline made it hard to fix these issues.

Instead I will create a set of examples which are much simpler and do not make use of the tank model. The tank model (and other non-free assets) will only be used for tech-demos which show the capabilities of the engine and are distributed in binary-only form.